refactor: 강의에 등록할 수 있는지 여부를 판단하는 로직을 session 객체에 위임
